Implemented the readDatabase Method
Wrote a method to run a query that returns all the records in the People
table from the JDBCDemo database on my localhost. Then prints the query
as a string above the results printed in a tabular format.

Added JavaDoc comments and removed the throws statement from the method
declaration. Eclipse suggested it before I wrote the try/catch block.
Handling exceptions with try/catch makes 'throws Exception' unnecessary
.