Added a close() Method
Added a method to close the connection and 'free' the resources from
JDBC instances. Also added a finally block to the main method, where the
new close method gets called. I also moved the kb.close(); statement to
the finally block. It makes more sense for it to be there.